IF-19: Refactoring
