+++ fusionfrontend/src/app/ecosystem/components/content/asset-type-template-create/asset-type-template-create-dialog/asset-type-template-create-dialog.component.html
