+++ fusionfrontend/src/app/ecosystem/components/content/asset-type-template-create/asset-type-template-create-dialog/asset-type-template-create-dialog.component.scss
@@ -13,9 +13,9 @@
  * under the License.
  */
 
-@import "base/typography";
-@import "abstract/variables";
-@import "components/buttons";
+@import "src/assets/sass/base/typography";
+@import "src/assets/sass/abstract/variables";
+@import "src/assets/sass/components/buttons";
 
 .wrapper {
   display: flex;
