+++ fusionfrontend/src/app/ecosystem/components/content/asset-type-template-create/asset-type-template-create-step-finished/asset-type-template-create-step-finished.component.scss
@@ -13,7 +13,7 @@
  * under the License.
  */
 
-@import '../asset-type-template-create/asset-type-template-create.component.scss';
+@import 'src/app/ecosystem/components/content/asset-type-template-create/asset-type-template-create-dialog/asset-type-template-create-dialog.component';
 
 .description {
   margin-top: 24px;
