+++ fusionfrontend/src/app/ecosystem/components/content/asset-type-template-create/asset-type-template-create-step-finished/asset-type-template-create-step-finished.component.spec.ts
