+++ fusionfrontend/src/app/ecosystem/components/content/asset-type-template-create/asset-type-template-create-step-four/asset-type-template-create-step-four.component.html
@@ -40,5 +40,6 @@ under the License.
 </div>
 <div class="footer">
   <button type="submit" class="if-button button-cancel" (click)="onBackToAttributes()">Back</button>
-  <button type="button" class="if-button button-confirm button-primary" (click)="onSave(5)">Save & Close Template</button>
+  <button type="button" class="if-button button-confirm button-primary" (click)="onSave()">Save & Close Template</button>
+  <button type="button" class="if-button button-confirm button-primary" (click)="onGoToPublish()">Publish</button>
 </div>
