+++ fusionfrontend/src/app/ecosystem/components/content/asset-type-template-create/asset-type-template-create-step-four/asset-type-template-create-step-four.component.scss
@@ -13,4 +13,4 @@
  * under the License.
  */
 
-@import '../asset-type-template-create/asset-type-template-create.component.scss';
+@import 'src/app/ecosystem/components/content/asset-type-template-create/asset-type-template-create-dialog/asset-type-template-create-dialog.component';
