+++ fusionfrontend/src/app/ecosystem/components/content/asset-type-template-create/asset-type-template-create-step-four/asset-type-template-create-step-four.component.spec.ts
