+++ fusionfrontend/src/app/ecosystem/components/content/asset-type-template-create/asset-type-template-create-step-one/asset-type-template-create-step-one.component.spec.ts
