+++ fusionfrontend/src/app/ecosystem/components/content/asset-type-template-create/asset-type-template-create-step-three/asset-type-template-create-step-three.component.spec.ts
