+++ fusionfrontend/src/app/ecosystem/components/content/asset-type-template-create/asset-type-template-create-step-two/asset-type-template-create-step-two.component.html
